requirements.txt
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7
Pillow==10.0.0 